+++ react-native/components/index.tsx
@@ -1,6 +1,0 @@
-export * from './PlayButtonView';
-export * from './ChannelView';
-export * from './CustomLiveVoiceView';
-export * from './ReconnectingView';
-export * from './ErrorView';
-export * from './AudioOutputSwitchView';