+++ react-native/src/components/DemoProps.ts
@@ -1,0 +1,14 @@
+/*
+ * This source file is part of the LiveVoice SDK project
+ *
+ * Copyright (c) LiveVoice GmbH
+ * Licensed under the MIT license
+ *
+ * See LICENSE for license information
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+export interface DemoProps {
+    isDarkMode: boolean
+}