+++ react-native/src/components/index.tsx
@@ -1,0 +1,12 @@
+/*
+ * This source file is part of the LiveVoice SDK project
+ *
+ * Copyright (c) LiveVoice GmbH
+ * Licensed under the MIT license
+ *
+ * See LICENSE for license information
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+export * from './CustomLiveVoiceView';