+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Platforms: [iOS](#ios) | [Android](#android) | [React Native](#react-native)
 
 ## iOS
 
+### 2.0.0 - June 3, 2026
+
+- Updated for LiveVoice 5.0 model changes and UI
+- Get derived AI channels per channel
+
 ### 1.0.1 - May 13, 2025
 
 - Bugfixes and optimizations
@@ -14,6 +19,11 @@ Platforms: [iOS](#ios) | [Android](#android) | [React Native](#react-native)
 
 ## Android
 
+### 2.0.0 - June 3, 2026
+
+- Updated for LiveVoice 5.0 model changes and UI
+- Get derived AI channels per channel
+
 ### 1.0.1 - May 13, 2025
 
 - Bugfixes and optimizations
@@ -23,6 +33,11 @@ Platforms: [iOS](#ios) | [Android](#android) | [React Native](#react-native)
 - Initial release
 
 ## React Native
+
+### 2.0.0 - June 3, 2026
+
+- Updated for LiveVoice 5.0 model changes and UI
+- Get derived AI channels per channel
 
 ### 1.0.3 - May 14, 2025